// src/services/api/generated/models/model-type-enum.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Test Management API
 * API for managing boards, tests, and execution.
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * * `POWER_EXTENSION` - Power Extension * `ETH_008_A` - Ethernet 008 Type A * `ETH_008_B` - Ethernet 008 Type B * `ETH_016` - Ethernet 016 * `CUSTOM` - Custom
 */

export const ModelTypeEnum = {
    POWER_EXTENSION: 'POWER_EXTENSION',
    ETH_008_A: 'ETH_008_A',
    ETH_008_B: 'ETH_008_B',
    ETH_016: 'ETH_016',
    CUSTOM: 'CUSTOM'
} as const;

export type ModelTypeEnum = typeof ModelTypeEnum[keyof typeof ModelTypeEnum];